Fix bug with VALUE function failing to handle empty string (#123)
* Fix bug with VALUE function failing to handle empty string, add corresponding tests.

* Update license header.
